+++ src/main/java/com/breakinblocks/plonk/common/tile/TilePlacedItems.java
@@ -7,8 +7,8 @@ import com.breakinblocks.plonk.common.util.ItemUtils;
 import com.breakinblocks.plonk.common.util.bound.Box;
 import com.breakinblocks.plonk.common.util.bound.BoxCollection;
 import com.google.common.collect.ImmutableMap;
-import net.minecraft.block.Blocks;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.fluid.IFluidState;
 import net.minecraft.inventory.ISidedInventory;
 import net.minecraft.inventory.ItemStackHelper;
 import net.minecraft.item.ItemStack;
@@ -118,7 +118,8 @@ public class TilePlacedItems extends TileEntity implements ISidedInventory, ITic
         // If empty tile
         if (last_not_empty == -1) {
             Objects.requireNonNull(this.world);
-            this.world.setBlockState(this.pos, Blocks.AIR.getDefaultState());
+            IFluidState fluidState = this.world.getFluidState(this.pos);
+            this.world.setBlockState(this.pos, fluidState.getBlockState());
             return true;
         }