+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [NEXT] - YYYY-MM-DD
+- Fix removing last item setting to air instead of using fluid state
+
 ## [7.0.1] - 2021-01-01
 - Add command to dump item stack render type data
 - Adjust render type heuristic: Signs, beds and shields now use item rendering